The method in the controller for generating the QR code is transferred to the service and registered in the DI container.
